+++ NEWS.md
@@ -87,10 +87,13 @@ grid_terrain(las, 1, tin(), keep_lowest = TRUE, full_raster = TRUE, use_class = 
 * Fix: In `segment_trees()` when using a raster-based algorithm, few points may be missclassified NAs at the edges of the point cloud instead of getting the correct tree ID found in the raster because of some edge effect.
 * Fix: `normalize_intensity()` was not working with a `LAScatalog`. See [#388](https://github.com/Jean-Romain/lidR/issues/388)
 * Fix: In `grid_*()` function when a `RasterLayer` is given as layout, the computation was performed for all the cells no matter if the extent of the loaded point cloud was much smaller than the raster. For big rasters this had for consequences to dramatically increase the wordload with useless computation and to saturate the RAM to a point that the computation was no longer doable.
+* Fix: In `track_sensor()` pulseID could be wrongly attributed for multi-beam sensors if the number of points very low. See [#388](https://github.com/Jean-Romain/lidR/issues/392)
+* Fix: In `track_sensor()` if `thin_pulses_with_time = 0` a single pulse was loaded with a `LAScatalog`. However it worked with a LAS.
 * Doc: The documentation of `point_metrics()` clarifies how the user-defined function is fed and in which order the points are sorted.
 * Doc: The argument `Wdegerated` in `grid_terrain()` and `normalize_height()` was misleading.  An wrong interpretation was that degeratated ground points were discared from the dataset. The documentation now clarrifies the text to avoid misinterpretation.
 * Doc: minor fixes and clarifications in the `LAScatalog-class` page of the manual.
 * Enhance: `plot_dtm3d()` now enables pan by default like `plot()` for `LAS` objects.
+* Enhance: `track_sensor()` throw a new warning is a swath in the point cloud does not produce any sensor location. This adresses [#391](https://github.com/Jean-Romain/lidR/issues/391).
 
 ## lidR v3.0.4 (Release date: 2020-10-08)